backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
docling>=2.0.0
langchain==0.2.6
langchain-google-genai==1.0.6
langchain-groq==0.1.6
supabase==2.5.0
pydantic-settings==2.3.0
python-dotenv==1.0.1
pytest==8.2.0
pytest-asyncio==0.23.7
ruff==0.4.8
